+++ R/chrono.subsets.R
@@ -117,7 +117,7 @@ chrono.subsets <- function(data, tree, method, time, model, inc.nodes = FALSE, F
     if(!missing(tree)) {
         check.class(tree, "phylo")
         ## tree must be dated
-        if(length(tree$root.time) != 1) stop("Tree must be a dated tree with a $root.time element.", call. = FALSE)
+        if(length(tree$root.time) != 1) stop("Tree must be a dated tree with a $root.time element.")
         ## tree.age_tree variable declaration
         tree.age_tree <- tree.age(tree)
     }
@@ -139,14 +139,14 @@ chrono.subsets <- function(data, tree, method, time, model, inc.nodes = FALSE, F
 
     ## If the tree is missing, the method can intake a star tree (i.e. no phylogeny)
     if(missing(tree)) {
-        if(missing(FADLAD)) stop("If no phylogeny is provided, all elements must be present in the FADLAD argument.", call. = FALSE)
-        if(method == "continuous") stop("If no phylogeny is provided, method must be \"discrete\".", call. = FALSE)
+        if(missing(FADLAD)) stop("If no phylogeny is provided, all elements must be present in the FADLAD argument.")
+        if(method == "continuous") stop("If no phylogeny is provided, method must be \"discrete\".")
         tree_was_missing <- TRUE
 
         ## Checking FADLAD disponibilities
         names_data <- rownames(data)
         ## All names must be present in the data
-        if(!all(names_data %in% rownames(FADLAD))) stop("If no phylogeny is provided, all elements must be present in the FADLAD argument.", call. = FALSE)
+        if(!all(names_data %in% rownames(FADLAD))) stop("If no phylogeny is provided, all elements must be present in the FADLAD argument.")
         ## Generating the star tree
         tree <- stree(nrow_data, tip.label = names_data)
         tree$root.time <- max(FADLAD)
@@ -170,7 +170,7 @@ chrono.subsets <- function(data, tree, method, time, model, inc.nodes = FALSE, F
     ## If time is a single value create the time vector by sampling evenly from just after the tree root time (1%) to the present
     if(length(time) == 1) {
         ## time must be at least three if discrete
-        if(time < 2) stop("Time must be greater or equal than 2.", call. = FALSE)
+        if(time < 2) stop("Time must be greater or equal than 2.")
 
         if(missing(tree)) {
             ## Set tmax
@@ -198,7 +198,7 @@ chrono.subsets <- function(data, tree, method, time, model, inc.nodes = FALSE, F
                 }
             } else {
                 if(t0 > max(tree.age_tree[,1]) || t0 < min(tree.age_tree[,1])) {
-                    stop("t0 is out of the tree age range.", call. = FALSE)
+                    stop("t0 is out of the tree age range.")
                 }
             }
         }
@@ -233,7 +233,7 @@ chrono.subsets <- function(data, tree, method, time, model, inc.nodes = FALSE, F
         ## else inc.nodes must be logical
         check.class(inc.nodes, "logical")
         if(tree_was_missing && inc.nodes) {
-            stop("If no phylogeny is provided, inc.nodes must be FALSE.", call. = FALSE)
+            stop("If no phylogeny is provided, inc.nodes must be FALSE.")
         }
     } else {
         ## Include nodes is mandatory
@@ -245,17 +245,17 @@ chrono.subsets <- function(data, tree, method, time, model, inc.nodes = FALSE, F
     if(inc.nodes != TRUE) {
         ## Check if at least all the data in the table are present in the tree
         if(any(is.na(match(rownames(data), tree$tip.label)))) {
-            stop("The labels in the matrix and in the tree do not match!\nUse clean.data() to match both tree and data or make sure whether nodes should be included or not (inc.nodes = FALSE by default).", call. = FALSE)
+            stop("The labels in the matrix and in the tree do not match!\nUse clean.data() to match both tree and data or make sure whether nodes should be included or not (inc.nodes = FALSE by default).")
         }
     } else {
         ## Check if the tree has node labels
         if(length(tree$node.label) != 0) {
             ## Check if the tree and the table are the same length
-            if(nrow_data != (Ntip_tree + Nnode(tree))) stop("The labels in the matrix and in the tree do not match!\nRemember to check the node labels in the tree and the matrix.", call. = FALSE)
+            if(nrow_data != (Ntip_tree + Nnode(tree))) stop("The labels in the matrix and in the tree do not match!\nRemember to check the node labels in the tree and the matrix.")
             ## Check if both nodes and tip labels match with the data rownames
-            if(any(is.na(c(rownames(data), c(tree$tip.label, tree$node.label))))) stop("The labels in the matrix and in the tree do not match!\nCheck especially the node labels in the tree and the matrix.", call. = FALSE)
+            if(any(is.na(c(rownames(data), c(tree$tip.label, tree$node.label))))) stop("The labels in the matrix and in the tree do not match!\nCheck especially the node labels in the tree and the matrix.")
         } else {
-            stop("The labels in the matrix and in the tree do not match!\nRemember to check the node labels in the tree and the matrix.", call. = FALSE)
+            stop("The labels in the matrix and in the tree do not match!\nRemember to check the node labels in the tree and the matrix.")
         }
     }
 
@@ -269,7 +269,7 @@ chrono.subsets <- function(data, tree, method, time, model, inc.nodes = FALSE, F
         check.class(FADLAD, "data.frame")
 
         if(!all(colnames(FADLAD) %in% c("FAD", "LAD"))) {
-            stop("FADLAD must be a data.frame with two columns being called respectively:\n\"FAD\" (First Apparition Datum) and \"LAD\" (Last Apparition Datum).", call. = FALSE)
+            stop("FADLAD must be a data.frame with two columns being called respectively:\n\"FAD\" (First Apparition Datum) and \"LAD\" (Last Apparition Datum).")
         } else {
             ## Check if FAD/LAD is in the right order (else reorder)
             if(colnames(FADLAD)[1] == "LAD") {