+++ R/custom.subsets.R
@@ -91,33 +91,33 @@ custom.subsets <- function(data, group) {
                 group_class <- "phylo"
                 ## Matching the tree and the data
                 if(any(is.na(match(group$tip.label, rownames(data))))) {
-                    stop("Some tips in the tree are not matching the data.\nSee ?clean.data for matching the tree and the data.", call. = FALSE)
+                    stop("Some tips in the tree are not matching the data.\nSee ?clean.data for matching the tree and the data.")
                 }
 
                 if(is.null(group$node.labe)) {
                     ## No nodes
                     if(nrow(data) != Ntip(group)) {
-                        stop("Some rows in the data are not matching the tree.\nSee ?clean.data for matching the tree and the data.", call. = FALSE)
+                        stop("Some rows in the data are not matching the tree.\nSee ?clean.data for matching the tree and the data.")
                     }
                     inc.nodes <- FALSE
                 } else {
                     ## Are the nodes relevant in the data
                     if(nrow(data) == (Ntip(group) + Nnode(group))) {
                         if(any(is.na(match(group$node.label, rownames(data))))) {
-                            stop("Some nodes in the tree are not matching the data.\nSee ?clean.data for matching the tree and the data.", call. = FALSE)
+                            stop("Some nodes in the tree are not matching the data.\nSee ?clean.data for matching the tree and the data.")
                         }
                         inc.nodes <- TRUE
                     } else {
                         if(nrow(data) == Ntip(group)) {
                             inc.nodes <- FALSE
                         } else {
-                            stop("Some rows in the data are not matching the tree.\nSee ?clean.data for matching the tree and the data.", call. = FALSE)
+                            stop("Some rows in the data are not matching the tree.\nSee ?clean.data for matching the tree and the data.")
                         }
                     }
                 }
 
             } else {
-                stop("group argument must be either a 'list', a 'matrix', a 'data.frame' or a 'phylo' object.", call. = FALSE)
+                stop("group argument must be either a 'list', a 'matrix', a 'data.frame' or a 'phylo' object.")
             }
         }
     }
@@ -129,10 +129,10 @@ custom.subsets <- function(data, group) {
         ## Using a data.frame or a matrix
 
         ## must have the same labels as data
-        if(!all( as.character(rownames(group)) %in% as.character(rownames(data)))) stop("Row names in data and group arguments don't match.", call. = FALSE)
+        if(!all( as.character(rownames(group)) %in% as.character(rownames(data)))) stop("Row names in data and group arguments don't match.")
 
         ## Checking if the groups have a list three elements
-        if(any(apply(group, 2, check.elements.data.frame))) stop("There must be at least three elements for each subset.", call. = FALSE)
+        if(any(apply(group, 2, check.elements.data.frame))) stop("There must be at least three elements for each subset.")
 
         ## Creating the subsets
         subsets_list <- unlist(apply(group, 2, split.elements.data.frame, data), recursive = FALSE)
@@ -184,16 +184,16 @@ custom.subsets <- function(data, group) {
         ## Cleaning groups
         if(all(unique(unlist(lapply(group[group_select], class))) %in% c("numeric", "integer"))) {
             ## The list must have the same columns as in the data
-            if(max(unlist(group[group_select])) > nrow(data)) stop("Row numbers in group don't match the row numbers in data.", call. = FALSE)
+            if(max(unlist(group[group_select])) > nrow(data)) stop("Row numbers in group don't match the row numbers in data.")
         } else {
             if(all(unique(unlist(lapply(group[group_select], class))) == "character")) {
-                if(!all( as.character(unlist(group[group_select])) %in% as.character(rownames(data)))) stop("Row names in data and group arguments don't match.", call. = FALSE)
+                if(!all( as.character(unlist(group[group_select])) %in% as.character(rownames(data)))) stop("Row names in data and group arguments don't match.")
                 
                 ## Convert the row names into row numbers
                 group <- lapply(group, convert.name.to.numbers, data)
 
             } else {
-                stop("group argument must be a list of row names or row numbers.", call. = FALSE)
+                stop("group argument must be a list of row names or row numbers.")
             }
         }