+++ R/plot.dispRity.R
@@ -202,7 +202,7 @@ plot.dispRity <- function(x, ..., type, quantiles = c(50, 95), cent.tend = media
                 }
                 ## Are quantiles proper proportions
                 if(any(quantiles < 0) | any(quantiles > 100)) {
-                    stop("quantiles(s) must be any value between 0 and 100.", call. = FALSE)
+                    stop("quantiles(s) must be any value between 0 and 100.")
                 }
                 quantiles_n <- length(quantiles)
 
@@ -210,7 +210,7 @@ plot.dispRity <- function(x, ..., type, quantiles = c(50, 95), cent.tend = media
                 check.class(cent.tend, "function")
                 ## The function must work
                 if(make.metric(cent.tend, silent = TRUE) != "level1") {
-                    stop("cent.tend argument must be a function that outputs a single numeric value.", call. = FALSE)
+                    stop("cent.tend argument must be a function that outputs a single numeric value.")
                 }
 
 
@@ -264,7 +264,7 @@ plot.dispRity <- function(x, ..., type, quantiles = c(50, 95), cent.tend = media
     ## must be class dispRity
     check.class(data, "dispRity")
     ## must have one element called dispRity
-    if(is.na(match("disparity", names(data)))) stop("Data must be a 'dispRity' object.", call. = FALSE)
+    if(is.na(match("disparity", names(data)))) stop("Data must be a 'dispRity' object.")
     ## Check if disparity is a value or a distribution
     is_distribution <- ifelse(length(data$disparity[[1]]$elements) != 1, TRUE, FALSE)
     ## Check the bootstraps
@@ -282,7 +282,7 @@ plot.dispRity <- function(x, ..., type, quantiles = c(50, 95), cent.tend = media
         }
         ## Are quantiles proper proportions
         if(any(quantiles < 0) | any(quantiles > 100)) {
-            stop("quantiles(s) must be any value between 0 and 100.", call. = FALSE)
+            stop("quantiles(s) must be any value between 0 and 100.")
         }
     }
 
@@ -291,7 +291,7 @@ plot.dispRity <- function(x, ..., type, quantiles = c(50, 95), cent.tend = media
     check.class(cent.tend, "function")
     ## The function must work
     if(make.metric(cent.tend, silent = TRUE) != "level1") {
-        stop("cent.tend argument must be a function that outputs a single numeric value.", call. = FALSE)
+        stop("cent.tend argument must be a function that outputs a single numeric value.")
     }
 
     ## type
@@ -386,7 +386,7 @@ plot.dispRity <- function(x, ..., type, quantiles = c(50, 95), cent.tend = media
         if(elements == FALSE) {
             check.length(ylab, 1, " must be a character string.")
         } else {
-            if(length(ylab) > 2) stop("ylab can have maximum of two elements.", call. = FALSE)
+            if(length(ylab) > 2) stop("ylab can have maximum of two elements.")
         }
     }