+++ NEWS.md
@@ -24,12 +24,20 @@ dispRity 0.2.0 (2016-04-01)
 [Next release] dispRity v1.1.1  (2018-05-03)
 =========================
 
+### NEW FEATURES
+
+  * S3 `print` method for objects of class `"dtt"` and `"dispRity"` (from `dtt.dispRity`).
+
 ### MINOR IMPROVEMENTS
 
+  * `dtt.dispRity` now allows alternative hypothesis specification (if `nsim > 0`).
   * `ellipse.volume` can now take an explicit eigen value vector (still unnecessary and estimated correctly for PCO and MDS)
   * Improved metric checking messages from `make.metric` when dealing with optional arguments
   * Removed cascade of warnings triggered by `plot.dispRity.dtt`
 
+
+<!--   * `dtt.dispRity` now works with any tree (including non-ultrametric ones and fossils only ones). -->
+<!--   * `dtt.dispRity` now works with time-slicing. -->
 
 [GitHub release] dispRity v1.1 *got CRAN* (2018-03-20)
 =========================